fixed bug related to multiple repo branches sharing the same db
git-svn-id: http://mysql-php-migrations.googlecode.com/svn/trunk@22 1f80a9a2-aebd-11dd-8f1c-4d87b713b184

+++ lib/helpers/list_helper.php
@@ -99,7 +99,7 @@ class MpmListHelper
         }
         $pdo->commit();
         $pdo->beginTransaction();
-        // remove migrations from the database which no longer have a corresponding file
+        // remove migrations from the database which no longer have a corresponding file and are not active yet
         try
         {
             $total_migrations = MpmListHelper::getTotalMigrations();
@@ -108,7 +108,7 @@ class MpmListHelper
             $file_timestamps = MpmListHelper::getTimestampArray($files);
             foreach ($db_list as $obj)
             {
-                if (!in_array($obj->timestamp, $file_timestamps))
+                if (!in_array($obj->timestamp, $file_timestamps) && $obj->active == 0)
                 {
                     $sql = "DELETE FROM `mpm_migrations` WHERE `id` = '{$obj->id}'";
                     $pdo->exec($sql);
